requirements.txt
fastapi
uvicorn
python-multipart
requests
psycopg2-binary
numpy
opencv-python-headless
torch
transformers
pillow
pandas